refactor: put entity and subscriber files in the root folder of the module
